User-provided-service supported.
To create a backingserviceinstance, use oc new-instance command.

+++ pkg/backingserviceinstance/controller/controller.go
@@ -43,10 +43,14 @@ func (c *BackingServiceInstanceController) Handle(bsi *backingserviceinstanceapi
 	//c.recorder.Eventf(bsi, "Debug", "bsi handler called.%s", bsi.Name)
 
 	changed := false
-
-	bs, err := c.Client.BackingServices("openshift").Get(bsi.Spec.BackingServiceName)
+	bs := &backingserviceapi.BackingService{}
+	if bsi.Annotations[backingserviceinstanceapi.UPS] != "true" {
+	bsp, err := c.Client.BackingServices("openshift").Get(bsi.Spec.BackingServiceName)
 	if err != nil {
 		return err
+	}else{
+		bs = bsp
+	}
 	}
 
 	switch bsi.Status.Phase {
@@ -115,10 +119,10 @@ func (c *BackingServiceInstanceController) Handle(bsi *backingserviceinstanceapi
 		svcinstance, err := servicebroker_create_instance(serviceinstance, bsInstanceID, servicebroker)
 		if err != nil {
 			result = err
-			c.recorder.Eventf(bsi,kapi.EventTypeWarning, "Provisioning", err.Error())
+			c.recorder.Eventf(bsi, kapi.EventTypeWarning, "Provisioning", err.Error())
 			break
 		} else {
-			c.recorder.Eventf(bsi,kapi.EventTypeNormal, "Provisioning", "bsi provisioning done, instanceid: %s", bsInstanceID)
+			c.recorder.Eventf(bsi, kapi.EventTypeNormal, "Provisioning", "bsi provisioning done, instanceid: %s", bsInstanceID)
 			glog.Infoln("bsi provisioning servicebroker_create_instance done, ", bsi.Name)
 		}
 
@@ -139,34 +143,59 @@ func (c *BackingServiceInstanceController) Handle(bsi *backingserviceinstanceapi
 	case backingserviceinstanceapi.BackingServiceInstancePhaseUnbound:
 		switch bsi.Status.Action {
 		case backingserviceinstanceapi.BackingServiceInstanceActionToDelete:
+			if bsi.Annotations[backingserviceinstanceapi.UPS] == "true" {
+				glog.Infoln(backingserviceinstanceapi.UPS, " bsi deleted ", bsi.Name)
 
-			if result = c.deleteInstance(bs, bsi); result == nil {
+				bsi.Status.Phase = backingserviceinstanceapi.BackingServiceInstancePhaseDeleted
+
+				bsi.Status.Action = ""
 				changed = true
+			} else {
+				if result = c.deleteInstance(bs, bsi); result == nil {
+					changed = true
+				}
 			}
-			c.recorder.Eventf(bsi,kapi.EventTypeNormal, "Deleting", "instance:%s [%v]", bsi.Name, changed)
+			c.recorder.Eventf(bsi, kapi.EventTypeNormal, "Deleting", "instance:%s [%v]", bsi.Name, changed)
 		case backingserviceinstanceapi.BackingServiceInstanceActionToBind:
-
 			dcname := c.get_deploymentconfig_name(bsi, backingserviceinstanceapi.BindDeploymentConfigBinding)
-			if result = c.bindInstance(dcname, bs, bsi); result == nil {
-				changed = true
+			if bsi.Annotations[backingserviceinstanceapi.UPS] == "true" {
+				if result = c.bindInstanceUPS(dcname, bsi); result == nil {
+					changed = true
+				}
+			}else {
+				if result = c.bindInstance(dcname, bs, bsi); result == nil {
+					changed = true
+				}
 			}
-			c.recorder.Eventf(bsi,kapi.EventTypeNormal, "Binding", "instance: %s, dc: %s [%v]", bsi.Name, dcname, changed)
-
+			c.recorder.Eventf(bsi, kapi.EventTypeNormal, "Binding", "instance: %s, dc: %s [%v]", bsi.Name, dcname, changed)
 		}
 	case backingserviceinstanceapi.BackingServiceInstancePhaseBound:
 		switch bsi.Status.Action {
 		case backingserviceinstanceapi.BackingServiceInstanceActionToUnbind:
+
 			dcname := c.get_deploymentconfig_name(bsi, backingserviceinstanceapi.BindDeploymentConfigUnbinding)
-			if result = c.unbindInstance(dcname, bs, bsi); result == nil {
-				changed = true
+			if bsi.Annotations[backingserviceinstanceapi.UPS] == "true"{
+				if result = c.unbindInstanceUPS(dcname, bsi); result == nil {
+					changed = true
+				}
+			}else {
+				if result = c.unbindInstance(dcname, bs, bsi); result == nil {
+					changed = true
+				}
 			}
-			c.recorder.Eventf(bsi,kapi.EventTypeNormal, "Unbinding", "instance: %s, dc: %s [%v]", bsi.Name, dcname, changed)
+			c.recorder.Eventf(bsi, kapi.EventTypeNormal, "Unbinding", "instance: %s, dc: %s [%v]", bsi.Name, dcname, changed)
 		case backingserviceinstanceapi.BackingServiceInstanceActionToBind:
 			dcname := c.get_deploymentconfig_name(bsi, backingserviceinstanceapi.BindDeploymentConfigBinding)
+			if bsi.Annotations[backingserviceinstanceapi.UPS] == "true" {
+				if result = c.bindInstanceUPS(dcname, bsi); result == nil {
+					changed = true
+				}
+			}else {
 			if result = c.bindInstance(dcname, bs, bsi); result == nil {
 				changed = true
 			}
-			c.recorder.Eventf(bsi,kapi.EventTypeNormal, "Binding", "instance: %s, dc: %s [%v]", bsi.Name, dcname, changed)
+			}
+			c.recorder.Eventf(bsi, kapi.EventTypeNormal, "Binding", "instance: %s, dc: %s [%v]", bsi.Name, dcname, changed)
 			/*
 				default:
 					return fmt.Errorf("action '%s' should never happen under status '%s'", bsi.Status.Action, bsi.Status.Phase)
@@ -187,7 +216,7 @@ func (c *BackingServiceInstanceController) Handle(bsi *backingserviceinstanceapi
 		*/
 
 		glog.Infoln("bsi controller error. ", err_msg)
-		c.recorder.Eventf(bsi,kapi.EventTypeWarning, "Error", err_msg)
+		c.recorder.Eventf(bsi, kapi.EventTypeWarning, "Error", err_msg)
 	}
 
 	if changed {
@@ -503,6 +532,13 @@ func (c *BackingServiceInstanceController) deploymentconfig_clear_envs(dc string
 	return c.deploymentconfig_modify_envs(dc, bsi, b, false)
 }
 
+
+func (c *BackingServiceInstanceController) deploymentconfig_inject_envs_ups(dc string, bsi *backingserviceinstanceapi.BackingServiceInstance, b *backingserviceinstanceapi.InstanceBinding) error {
+	return c.deploymentconfig_modify_envs_ups(dc, bsi, b, true)
+}
+func (c *BackingServiceInstanceController) deploymentconfig_clear_envs_ups(dc string, bsi *backingserviceinstanceapi.BackingServiceInstance, b *backingserviceinstanceapi.InstanceBinding) error {
+	return c.deploymentconfig_modify_envs_ups(dc, bsi, b, false)
+}
 // return exists or not
 func env_get(envs []kapi.EnvVar, envName string) (bool, string) {
 	for i := len(envs) - 1; i >= 0; i-- {
@@ -551,6 +587,59 @@ func env_unset(envs []kapi.EnvVar, envName string) (bool, []kapi.EnvVar) {
 	return index < n, envs[:index]
 }
 
+func (c *BackingServiceInstanceController) deploymentconfig_modify_envs_ups(dcname string, bsi *backingserviceinstanceapi.BackingServiceInstance, binding *backingserviceinstanceapi.InstanceBinding, toInject bool) error {
+	dc, err := c.Client.DeploymentConfigs(bsi.Namespace).Get(dcname)
+	if err != nil {
+		return err
+	}
+
+	if dc.Spec.Template == nil {
+		return nil
+	}
+
+	env_prefix := deploymentconfig_env_prefix(bsi.Spec.BackingServiceName, bsi.Name)
+	containers := dc.Spec.Template.Spec.Containers
+	if toInject {
+
+
+		vsp := &VcapServiceParameters{
+				Name:        bsi.Name,
+				Label:       "",
+				Plan:        backingserviceinstanceapi.UPS,
+				Credentials: binding.Credentials,
+			}
+
+
+		for i := range containers {
+			for k, v := range binding.Credentials {
+				_, containers[i].Env = env_set(containers[i].Env, deploymentconfig_env_name(env_prefix, k), v)
+			}
+
+			if vsp != nil {
+				_, containers[i].Env = modifyVcapServicesEnvNameEnv(containers[i].Env, backingserviceinstanceapi.UPS, vsp, "")
+			}
+		}
+		dc.Annotations["backingservice.instance/"+bsi.Name] = "bound"
+	} else {
+		for i := range containers {
+			for k := range binding.Credentials {
+				_, containers[i].Env = env_unset(containers[i].Env, deploymentconfig_env_name(env_prefix, k))
+			}
+
+			_, containers[i].Env = modifyVcapServicesEnvNameEnv(containers[i].Env, backingserviceinstanceapi.UPS, nil, bsi.Name)
+		}
+		delete(dc.Annotations, "backingservice.instance/"+bsi.Name)
+	}
+
+	if _, err := c.Client.DeploymentConfigs(bsi.Namespace).Update(dc); err != nil {
+		return err
+	}
+
+	c.deploymentconfig_print_envs(bsi.Namespace, binding)
+	return nil
+}
+
+
 func (c *BackingServiceInstanceController) deploymentconfig_modify_envs(dcname string, bsi *backingserviceinstanceapi.BackingServiceInstance, binding *backingserviceinstanceapi.InstanceBinding, toInject bool) error {
 	dc, err := c.Client.DeploymentConfigs(bsi.Namespace).Get(dcname)
 	if err != nil {
@@ -561,17 +650,18 @@ func (c *BackingServiceInstanceController) deploymentconfig_modify_envs(dcname s
 		return nil
 	}
 
-	bs, err := c.Client.BackingServices("openshift").Get(bsi.Spec.BackingServiceName)
-	if err != nil {
-		return err
-	}
-
-	var plan = (*backingserviceapi.ServicePlan)(nil)
-	for k := range bs.Spec.Plans {
-		if bsi.Spec.BackingServicePlanGuid == bs.Spec.Plans[k].Id {
-			plan = &(bs.Spec.Plans[k])
+		bs, err := c.Client.BackingServices("openshift").Get(bsi.Spec.BackingServiceName)
+		if err != nil {
+			return err
 		}
-	}
+
+		var plan = (*backingserviceapi.ServicePlan)(nil)
+		for k := range bs.Spec.Plans {
+			if bsi.Spec.BackingServicePlanGuid == bs.Spec.Plans[k].Id {
+				plan = &(bs.Spec.Plans[k])
+			}
+		}
+
 
 	env_prefix := deploymentconfig_env_prefix(bsi.Spec.BackingServiceName, bsi.Name)
 	containers := dc.Spec.Template.Spec.Containers
@@ -596,6 +686,7 @@ func (c *BackingServiceInstanceController) deploymentconfig_modify_envs(dcname s
 				_, containers[i].Env = modifyVcapServicesEnvNameEnv(containers[i].Env, bs.Name, vsp, "")
 			}
 		}
+		dc.Annotations["backingservice.instance/"+bsi.Name] = "bound"
 	} else {
 		for i := range containers {
 			for k := range binding.Credentials {
@@ -604,6 +695,7 @@ func (c *BackingServiceInstanceController) deploymentconfig_modify_envs(dcname s
 
 			_, containers[i].Env = modifyVcapServicesEnvNameEnv(containers[i].Env, bs.Name, nil, bsi.Name)
 		}
+		delete(dc.Annotations, "backingservice.instance/"+bsi.Name)
 	}
 
 	if _, err := c.Client.DeploymentConfigs(bsi.Namespace).Update(dc); err != nil {
@@ -754,6 +846,48 @@ func (c *BackingServiceInstanceController) deleteInstance(bs *backingserviceapi.
 
 }
 
+
+func (c *BackingServiceInstanceController) bindInstanceUPS(dc string,  bsi *backingserviceinstanceapi.BackingServiceInstance) (err error) {
+	glog.Infoln(backingserviceinstanceapi.UPS, "bsi to bind ", bsi.Name, " and ", dc)
+
+	instanceBinding := backingserviceinstanceapi.InstanceBinding{}
+	now := unversioned.Now()
+	instanceBinding.BoundTime = &now //&unversioned.Now()
+	instanceBinding.BindUuid = backingserviceinstanceapi.UPS
+	instanceBinding.BindDeploymentConfig = dc
+	instanceBinding.Credentials = bsi.Spec.Credentials
+
+
+	/*
+	instanceBinding.Credentials = make(map[string]string)
+	for k, v := range bsi.Spec.Credentials{
+		instanceBinding.Credentials[k] = v
+	}
+	*/
+
+	glog.Infoln("deploymentconfig_inject_envs")
+
+	err = c.deploymentconfig_inject_envs_ups(dc, bsi, &instanceBinding)
+	if err != nil {
+		return err
+	} else {
+		bsi.Spec.Binding = append(bsi.Spec.Binding, instanceBinding)
+	}
+
+	glog.Infoln("bsi bound. ", bsi.Name)
+
+	bsi.Spec.Bound += 1
+
+	bsi.Status.Phase = backingserviceinstanceapi.BackingServiceInstancePhaseBound
+	bsi.Annotations[dc] = backingserviceinstanceapi.BindDeploymentConfigBound
+
+	bsi.Status.Action = "" //remove_action_word(bsi.Status.Action, backingserviceinstanceapi.BackingServiceInstanceActionToBind)
+	return
+
+
+}
+
+
 func (c *BackingServiceInstanceController) bindInstance(dc string, bs *backingserviceapi.BackingService, bsi *backingserviceinstanceapi.BackingServiceInstance) (result error) {
 	glog.Infoln("bsi to bind ", bsi.Name, " and ", dc)
 
@@ -813,6 +947,52 @@ func (c *BackingServiceInstanceController) bindInstance(dc string, bs *backingse
 	bsi.Annotations[dc] = backingserviceinstanceapi.BindDeploymentConfigBound
 
 	bsi.Status.Action = "" //remove_action_word(bsi.Status.Action, backingserviceinstanceapi.BackingServiceInstanceActionToBind)
+	return
+
+}
+
+
+func (c *BackingServiceInstanceController) unbindInstanceUPS(dc string, bsi *backingserviceinstanceapi.BackingServiceInstance) (err error) {
+
+	glog.Infoln(backingserviceinstanceapi.UPS," bsi to unbind ", bsi.Name)
+
+
+	for idx, b := range bsi.Spec.Binding {
+		if b.BindDeploymentConfig == dc {
+
+			glog.Infoln("deploymentconfig_clear_envs")
+			err = c.deploymentconfig_clear_envs_ups(dc, bsi, &b)
+			if err != nil {
+				return err
+			} else {
+				bsi.Spec.Binding = append(bsi.Spec.Binding[:idx], bsi.Spec.Binding[idx+1:]...)
+				delete(bsi.Annotations, dc)
+			}
+			break
+		}
+	}
+	/*
+		err = c.deploymentconfig_clear_envs(bsi.Namespace, dc)
+		if err != nil {
+			return err
+		}
+	*/
+	glog.Infoln("bsi is unbound ", bsi.Name)
+	/*
+		delete(bsi.Annotations, dc)
+		bsi.Spec.BindDeploymentConfig = ""
+		bsi.Spec.Credentials = nil
+		bsi.Spec.BoundTime = nil
+		bsi.Spec.BindUuid = ""
+		bsi.Spec.Bound = false
+	*/
+	bsi.Spec.Bound -= 1
+	if bsi.Spec.Bound == 0 {
+		bsi.Status.Phase = backingserviceinstanceapi.BackingServiceInstancePhaseUnbound
+	}
+
+	bsi.Status.Action = "" //remove_action_word(bsi.Status.Action, backingserviceinstanceapi.BackingServiceInstanceActionToUnbind)
+
 	return
 
 }
